app/requirements.txt
Flask
flask-mysql
simplejson
python-dateutil